+++ src/Core/TundraProtocolModule/MsgAssetDeleted.h
@@ -45,9 +45,9 @@ struct MsgAssetDeleted
 
 	inline void SerializeTo(kNet::DataSerializer &dst) const
 	{
-		dst.Add<u8>(assetRef.size());
+		dst.Add<u8>((u8)assetRef.size());
 		if (assetRef.size() > 0)
-			dst.AddArray<s8>(&assetRef[0], assetRef.size());
+			dst.AddArray<s8>(&assetRef[0], (u32)assetRef.size());
 	}
 
 	inline void DeserializeFrom(kNet::DataDeserializer &src)
